pyproject.toml
[tool.poetry]
name = "meetmate"
version = "0.1.0"
description = ""
authors = ["S1avv <code.launchx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
aiohttp = "^3.9.3"
python-tgpt = "^0.5.3"
g4f = "^0.2.2.0"
aiosqlite = "^0.20.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
